+++ ElvUI_MerathilisUI/Core/Changelog/7.30.lua
@@ -4,9 +4,13 @@ MER.Changelog[730] = {
 	RELEASE_DATE = "TBD",
 	FIXES = {
 		"[Skins]: Fixed the Details Embed Width and Height options doing nothing",
+		"[BuffReminder]: Fixed group buff coverage checks failing when aura data was masked as a secret value",
+		"[Armory]: Fixed a tonumber error when looking up gem IDs in the Socket Panel",
 	},
 	NEW = {
 		"[Skins]: Added a Number of Windows option to the Details Embed system",
 	},
-	IMPROVEMENTS = {},
+	IMPROVEMENTS = {
+		"[Skins]: Updated the Weekly Rewards skin",
+	},
 }