+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ### Changes
 
 -   [Skins]: Fixed the Details Embed Width and Height options doing nothing
+-   [BuffReminder]: Fixed group buff coverage checks failing when aura data was masked as a secret value
+-   [Armory]: Fixed a tonumber error when looking up gem IDs in the Socket Panel
 -   [Skins]: Added a Number of Windows option to the Details Embed system
+-   [Skins]: Updated the Weekly Rewards skin